So, update/accident for the Gamescene please lmk if something happens.
I really don't quite know what I means but I think it might have recorded me accidently deleting the Gamescene and CTRL+Z cause of a slip of the finger sorry. But the main thing I did is what as follows is that I was just trying to adjust some the assets in the scene. plus with what I did last with the whole player setup for the main game scene.